Separate demographic variables into separate files
