+++ openfisca_us/variables/assets/spm_unit_assets.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_assets(Variable):
+    value_type = float
+    entity = SPMUnit
+    label = u"SPM unit assets"
+    definition_period = YEAR
+    unit = "currency-USD"
