+++ openfisca_us/variables/demographic/age/is_child.py
@@ -1,0 +1,11 @@
+from openfisca_us.model_api import *
+
+
+class is_child(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is a child"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        return person("age", period) < 18
