+++ openfisca_us/variables/demographic/age/is_senior.py
@@ -1,0 +1,11 @@
+from openfisca_us.model_api import *
+
+
+class is_senior(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is a senior"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        return person("age", period) >= 65
