+++ openfisca_us/variables/demographic/age/is_wa_adult.py
@@ -1,0 +1,12 @@
+from openfisca_us.model_api import *
+
+
+class is_wa_adult(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is a working-age adult"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        age = person("age", period)
+        return (age >= 18) & (age < 65)
