+++ openfisca_us/variables/demographic/geographic/state_code_str.py
@@ -1,0 +1,12 @@
+from openfisca_us.model_api import *
+
+
+class state_code_str(Variable):
+    value_type = str
+    entity = Household
+    label = "State code (string)"
+    documentation = "State code variable, stored as a string"
+    definition_period = YEAR
+
+    def formula(household, period, parameters):
+        return household("state_code", period).decode_to_str()
