+++ openfisca_us/variables/demographic/identifiers/household_id.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class household_id(Variable):
+    value_type = float
+    entity = Household
+    label = u"Unique reference for this household"
+    definition_period = ETERNITY
