+++ openfisca_us/variables/demographic/identifiers/person_family_id.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class person_family_id(Variable):
+    value_type = int
+    entity = Person
+    label = u"Unique reference for the family of this person"
+    definition_period = ETERNITY
