+++ openfisca_us/variables/demographic/identifiers/person_household_id.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class person_household_id(Variable):
+    value_type = int
+    entity = Person
+    label = u"Unique reference for the household of this person"
+    definition_period = ETERNITY
