+++ openfisca_us/variables/demographic/identifiers/tax_unit_id.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class tax_unit_id(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = u"Unique reference for this tax unit"
+    definition_period = ETERNITY
