+++ openfisca_us/variables/demographic/person/is_citizen.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class is_citizen(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is a U.S. citizen"
+    definition_period = YEAR
