+++ openfisca_us/variables/demographic/person/is_in_school.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class is_in_school(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is currently in an education institution"
+    definition_period = YEAR
