+++ openfisca_us/variables/demographic/spm_unit/spm_unit_size.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class spm_unit_size(Variable):
+    value_type = int
+    entity = SPMUnit
+    label = u"SPM unit size"
+    definition_period = YEAR
