+++ openfisca_us/variables/demographic/tax_unit/MARSType.py
@@ -1,0 +1,9 @@
+from openfisca_us.model_api import *
+
+
+class MARSType(Enum):
+    SINGLE = "Single"
+    JOINT = "Joint"
+    SEPARATE = "Separate"
+    HOUSEHOLD_HEAD = "Head of household"
+    WIDOW = "Widow(er)"
