+++ openfisca_us/variables/demographic/tax_unit/is_tax_unit_dependent.py
@@ -1,0 +1,14 @@
+from openfisca_us.model_api import *
+
+
+class is_tax_unit_dependent(Variable):
+    value_type = bool
+    entity = Person
+    label = u"Is a dependent in the tax unit"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        return not_(
+            person("is_tax_unit_head", period)
+            + person("is_tax_unit_spouse", period)
+        )
