+++ openfisca_us/variables/demographic/tax_unit/is_tax_unit_spouse.py
@@ -1,0 +1,12 @@
+from openfisca_us.model_api import *
+
+
+class is_tax_unit_spouse(Variable):
+    value_type = bool
+    entity = Person
+    label = "Spouse of tax unit"
+    definition_period = YEAR
+
+    def formula(person, period, parameters):
+        # Use order of input (second)
+        return person.tax_unit.members_position == 1
