+++ openfisca_us/variables/demographic/weights/household_weight.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class household_weight(Variable):
+    value_type = float
+    entity = Household
+    label = u"Household weight"
+    definition_period = YEAR
