+++ openfisca_us/variables/demographic/weights/person_weight.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class person_weight(Variable):
+    value_type = float
+    entity = Person
+    label = u"Person weight"
+    definition_period = YEAR
