+++ openfisca_us/variables/demographic/weights/tax_unit_weight.py
@@ -1,0 +1,8 @@
+from openfisca_us.model_api import *
+
+
+class tax_unit_weight(Variable):
+    value_type = float
+    entity = TaxUnit
+    label = u"Tax unit weight"
+    definition_period = YEAR
