+++ openfisca_us/variables/irs/inputs.py
@@ -1,7 +1,7 @@
 from openfisca_core.model_api import *
 from openfisca_us.entities import *
 from openfisca_us.tools.general import *
-from openfisca_us.variables.demographic.tax_unit import MARSType
+from openfisca_us.variables.demographic.tax_unit.MARSType import MARSType
 
 
 class dsi(Variable):
